:fire: modularize primitives
